requirements.txt
agate==1.6.3
aiofiles==0.8.0
aiohttp==3.8.1
aioresponses==0.7.3
aiosignal==1.2.0
alembic==1.7.7
amqp==5.1.0
anyio==3.5.0
apache-airflow==2.2.5
apache-airflow-providers-celery==2.1.0
apache-airflow-providers-ftp==2.1.0
apache-airflow-providers-http==2.1.0
apache-airflow-providers-imap==2.2.1
apache-airflow-providers-slack==4.2.3
apache-airflow-providers-sqlite==2.1.1
apispec==3.3.2
argcomplete==1.12.3
asttokens==2.0.5
async-timeout==4.0.2
asyncio==3.4.3
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.3.0
blinker==1.4
cachelib==0.6.0
cattrs==1.0.0
celery==5.1.2
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.0
coverage==6.3.2
croniter==1.3.4
cryptography==3.4.8
cycler==0.11.0
dbt-core==1.0.4
dbt-extractor==0.4.1
dbt-postgres==1.0.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docker==5.0.3
docutils==0.16
email-validator==1.1.3
executing==0.8.3
flake8==4.0.1
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==1.0.0
fonttools==4.32.0
freezegun==1.2.1
frozenlist==1.3.0
future==0.18.2
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hologram==0.0.14
httpcore==0.14.7
httpx==0.22.0
humanize==3.14.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.2.0
iso8601==1.0.2
isodate==0.6.1
isort==5.10.1
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jsonschema==3.2.0
kiwisolver==1.4.2
kombu==5.1.0
lazy-object-proxy==1.4.3
leather==0.3.4
lockfile==0.12.2
Logbook==1.5.3
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mashumaro==2.9
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
minimal-snowplow-tracker==0.0.2
msgpack==1.0.3
multidict==6.0.2
mypy==0.942
mypy-extensions==0.4.3
networkx==2.7.1
numpy==1.22.3
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
packaging==21.3
pandas==1.2.5
pandas-stubs==1.2.0.54
parsedatetime==2.4
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyfakefs==4.5.6
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==7.1.1
pytest-aiofiles==0.2.0
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest-httpx==0.20.0
pytest-mock==3.7.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytimeparse==1.1.8
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
rfc3986==1.5.0
rich==12.0.1
sendgrid==6.9.7
setproctitle==1.2.2
six==1.16.0
slack-sdk==3.15.2
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
sqlalchemy-stubs==0.4
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
stack-data==0.2.0
starkbank-ecdsa==2.0.3
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
types-aiofiles==0.8.5
types-all==1.0.0
types-annoy==1.17.5
types-atomicwrites==1.4.3
types-backports==0.1.3
types-backports-abc==0.5.2
types-beautifulsoup4==4.10.18
types-bleach==4.1.5
types-boto==2.49.10
types-cachetools==5.0.0
types-certifi==2021.10.8.1
types-characteristic==14.3.3
types-chardet==4.0.3
types-click==7.1.8
types-click-spinner==0.1.10
types-colorama==0.4.9
types-contextvars==2.4.4
types-croniter==1.0.8
types-cryptography==3.3.18
types-dataclasses==0.6.4
types-dateparser==1.1.0
types-DateTimeRange==1.2.4
types-decorator==5.1.4
types-Deprecated==1.2.5
types-docopt==0.6.6
types-docutils==0.18.0
types-emoji==1.2.7
types-enum34==1.1.8
types-fb303==1.0.0
types-filelock==3.2.5
types-first==2.0.2
types-Flask==1.1.6
types-freezegun==1.1.7
types-frozendict==2.0.7
types-futures==3.3.8
types-geoip2==3.0.0
types-ipaddress==1.0.8
types-itsdangerous==1.1.6
types-JACK-Client==0.5.4
types-Jinja2==2.11.9
types-kazoo==0.1.3
types-Markdown==3.3.12
types-MarkupSafe==1.1.10
types-maxminddb==1.5.0
types-mock==4.0.12
types-mypy-extensions==0.4.14
types-nmap==0.1.6
types-openssl-python==0.1.3
types-orjson==3.6.2
types-paramiko==2.8.17
types-pathlib2==2.3.0
types-Pillow==9.0.8
types-pkg-resources==0.1.3
types-polib==1.1.11
types-protobuf==3.19.15
types-psycopg2==2.9.9
types-pyaudio==0.2.11
types-pycurl==7.44.7
types-pyfarmhash==0.2.3
types-PyJWT==1.7.1
types-pymssql==2.1.0
types-PyMySQL==1.0.14
types-pyRFC3339==1.1.1
types-pysftp==0.2.12
types-python-dateutil==2.8.10
types-python-gflags==3.1.4
types-python-slugify==5.0.3
types-pytz==2021.3.6
types-pyvmomi==7.0.5
types-PyYAML==6.0.5
types-redis==4.1.18
types-requests==2.27.15
types-retry==0.9.5
types-Routes==2.5.0
types-scribe==2.0.0
types-simplejson==3.17.4
types-singledispatch==3.7.3
types-six==1.16.12
types-tabulate==0.8.6
types-termcolor==1.1.3
types-toml==0.10.4
types-tornado==5.1.1
types-typed-ast==1.5.2
types-tzlocal==4.1.0
types-ujson==4.2.1
types-urllib3==1.26.11
types-waitress==2.0.6
types-Werkzeug==1.0.9
types-xxhash==2.0.3
typing-extensions==4.1.1
unicode==2.8
unicodecsv==0.14.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.1
Werkzeug==1.0.1
wrapt==1.14.0
WTForms==2.3.3
yarl==1.7.2
zipp==3.6.0
